atcoder/200-299/c/abc267_c.cpp

+++ atcoder/200-299/c/abc267_c.cpp
@@ -11,16 +11,16 @@ int main() {
 
     ll n, m, cur_product = 0, max_product = 0;
     cin >> n >> m;
-    int left = 1, right = m;
+    ll left = 1, right = m;
 
 
     vector<ll> vc(n + 1);
     for (ll i = 1; i <= n; ++i) cin >> vc[i];
-    for (int i = left; i <= right; ++i) cur_product += (i - left + 1) * vc[i];
+    for (ll i = left; i <= right; ++i) cur_product += (i - left + 1) * vc[i];
     max_product = cur_product;
 
-    for (int i = m + 1; i <= n; ++i) {
-        for (int j = i - m; j <= i - 1; ++j) cur_product -= vc[j];
+    for (ll i = m + 1; i <= n; ++i) {
+        for (ll j = i - m; j <= i - 1; ++j) cur_product -= vc[j];
         cur_product += m * vc[i];
         max_product = max(max_product, cur_product);
     }
